update the tests
